prompt_gen/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
openai==0.28
psycopg2-binary
passlib[bcrypt]
python-jose
google-auth
firebase-admin
jinja2
pytz
python-dotenv
